reorganize with_lock decorator
